Adding new liveries for a320neo using this map
This is the plain map of the a320neo. I've added instruction in it too you may want to use this for making liveries. I've also added the Vistara(texture_2.jpg) and Spirit airlines(texture_21.jpg) liveries incase you may want to refer. Once you are done with the livery submit it to me by making an issue in here https://github.com/Spice9/Geofs-Multiliveries/issues.